requirements.txt
beautifulsoup4
lxml
PyYAML
requests
selenium
undetected-chromedriver
playwright
yt-dlp
